style fixes take 3!!!

+++ src/search/evaluation/psts.rs
@@ -2,7 +2,7 @@ use std::ops::{Neg};
 use super::super::eval::{*};
 
 #[rustfmt::skip]
-const MG_PAWN_TABLE: [i16; 64] = [
+const MG_PAWN_TABLE: [i32; 64] = [
     0,   0,   0,   0,   0,   0,  0,   0,
     98, 134,  61,  95,  68, 126, 34, -11,
     -6,   7,  26,  31,  65,  56, 25, -20,
@@ -14,7 +14,7 @@ const MG_PAWN_TABLE: [i16; 64] = [
 ];
 
 #[rustfmt::skip]
-const EG_PAWN_TABLE: [i16; 64] = [
+const EG_PAWN_TABLE: [i32; 64] = [
     0,   0,   0,   0,   0,   0,   0,   0,
     178, 173, 158, 134, 147, 132, 165, 187,
     94, 100,  85,  67,  56,  53,  82,  84,
@@ -26,7 +26,7 @@ const EG_PAWN_TABLE: [i16; 64] = [
 ];
 
 #[rustfmt::skip]
-const MG_KNIGHT_TABLE: [i16; 64] = [
+const MG_KNIGHT_TABLE: [i32; 64] = [
     -167, -89, -34, -49,  61, -97, -15, -107,
     -73, -41,  72,  36,  23,  62,   7,  -17,
     -47,  60,  37,  65,  84, 129,  73,   44,
@@ -38,7 +38,7 @@ const MG_KNIGHT_TABLE: [i16; 64] = [
 ];
 
 #[rustfmt::skip]
-const EG_KNIGHT_TABLE: [i16; 64] = [
+const EG_KNIGHT_TABLE: [i32; 64] = [
     -58, -38, -13, -28, -31, -27, -63, -99,
     -25,  -8, -25,  -2,  -9, -25, -24, -52,
     -24, -20,  10,   9,  -1,  -9, -19, -41,
@@ -50,7 +50,7 @@ const EG_KNIGHT_TABLE: [i16; 64] = [
 ];
 
 #[rustfmt::skip]
-const MG_BISHOP_TABLE: [i16; 64] = [
+const MG_BISHOP_TABLE: [i32; 64] = [
     -29,   4, -82, -37, -25, -42,   7,  -8,
     -26,  16, -18, -13,  30,  59,  18, -47,
     -16,  37,  43,  40,  35,  50,  37,  -2,
@@ -62,7 +62,7 @@ const MG_BISHOP_TABLE: [i16; 64] = [
 ];
 
 #[rustfmt::skip]
-const EG_BISHOP_TABLE: [i16; 64] = [
+const EG_BISHOP_TABLE: [i32; 64] = [
     -14, -21, -11,  -8, -7,  -9, -17, -24,
     -8,  -4,   7, -12, -3, -13,  -4, -14,
     2,  -8,   0,  -1, -2,   6,   0,   4,
@@ -74,7 +74,7 @@ const EG_BISHOP_TABLE: [i16; 64] = [
 ];
 
 #[rustfmt::skip]
-const MG_ROOK_TABLE: [i16; 64] = [
+const MG_ROOK_TABLE: [i32; 64] = [
     32,  42,  32,  51, 63,  9,  31,  43,
     27,  32,  58,  62, 80, 67,  26,  44,
     -5,  19,  26,  36, 17, 45,  61,  16,
@@ -86,7 +86,7 @@ const MG_ROOK_TABLE: [i16; 64] = [
 ];
 
 #[rustfmt::skip]
-const EG_ROOK_TABLE: [i16; 64] = [
+const EG_ROOK_TABLE: [i32; 64] = [
     13, 10, 18, 15, 12,  12,   8,   5,
     11, 13, 13, 11, -3,   3,   8,   3,
     7,  7,  7,  5,  4,  -3,  -5,  -3,
@@ -98,7 +98,7 @@ const EG_ROOK_TABLE: [i16; 64] = [
 ];
 
 #[rustfmt::skip]
-const MG_QUEEN_TABLE: [i16; 64] = [
+const MG_QUEEN_TABLE: [i32; 64] = [
     -28,   0,  29,  12,  59,  44,  43,  45,
     -24, -39,  -5,   1, -16,  57,  28,  54,
     -13, -17,   7,   8,  29,  56,  47,  57,
@@ -110,7 +110,7 @@ const MG_QUEEN_TABLE: [i16; 64] = [
 ];
 
 #[rustfmt::skip]
-const EG_QUEEN_TABLE: [i16; 64] = [
+const EG_QUEEN_TABLE: [i32; 64] = [
     -9,  22,  22,  27,  27,  19,  10,  20,
     -17,  20,  32,  41,  58,  25,  30,   0,
     -20,   6,   9,  49,  47,  35,  19,   9,
@@ -122,7 +122,7 @@ const EG_QUEEN_TABLE: [i16; 64] = [
 ];
 
 #[rustfmt::skip]
-const MG_KING_TABLE: [i16; 64] = [
+const MG_KING_TABLE: [i32; 64] = [
     -65,  23,  16, -15, -56, -34,   2,  13,
     29,  -1, -20,  -7,  -8,  -4, -38, -29,
     -9,  24,   2, -16, -20,   6,  22, -22,
@@ -134,7 +134,7 @@ const MG_KING_TABLE: [i16; 64] = [
 ];
 
 #[rustfmt::skip]
-const EG_KING_TABLE: [i16; 64] = [
+const EG_KING_TABLE: [i32; 64] = [
     -74, -35, -18, -18, -11,  15,   4, -17,
     -12,  17,  14,  17,  17,  38,  23,  11,
     10,  17,  23,  15,  20,  45,  44,  13,
@@ -145,7 +145,7 @@ const EG_KING_TABLE: [i16; 64] = [
     -53, -34, -21, -11, -28, -14, -24, -43
 ];
 
-const MG_PESTO_TABLE: [[i16; 64]; 6] = [
+const MG_PESTO_TABLE: [[i32; 64]; 6] = [
     MG_PAWN_TABLE,
     MG_KNIGHT_TABLE,
     MG_BISHOP_TABLE,
@@ -154,7 +154,7 @@ const MG_PESTO_TABLE: [[i16; 64]; 6] = [
     MG_KING_TABLE
 ];
 
-const EG_PESTO_TABLE: [[i16; 64]; 6] = [
+const EG_PESTO_TABLE: [[i32; 64]; 6] = [
     EG_PAWN_TABLE,
     EG_KNIGHT_TABLE,
     EG_BISHOP_TABLE,
@@ -164,8 +164,8 @@ const EG_PESTO_TABLE: [[i16; 64]; 6] = [
 ];
 
 pub struct PieceSquareTable {
-    pub mg_table: [[i16; 64]; 12],
-    pub eg_table: [[i16; 64]; 12],
+    pub mg_table: [[i32; 64]; 12],
+    pub eg_table: [[i32; 64]; 12],
 }
 
 impl PieceSquareTable {
